backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
openai==0.27.8
python-dotenv==1.0.0
pydantic>=2.0.0 